tsk-iahrh5: split MessagesApp monolith into MessagesApp/ directory + MobileMessages
Mirror the StoreApp structural pattern for taOStalk so the chat app is
not a single 2834-line file anymore and gains a focused mobile surface
the way the Store app has MobileStore.

Move (not rewrite):
- desktop/src/apps/MessagesApp.tsx -> desktop/src/apps/MessagesApp/index.tsx
  with all relative imports rewritten (./chat/... -> ../chat/...,
  ./MessagesApp.* -> ../MessagesApp.*).
- desktop/src/apps/MessagesApp.tsx becomes a 5-line re-export shim so the
  existing '@/apps/MessagesApp' and './MessagesApp' import paths keep
  resolving for every call site and test.

Mobile surface (extract, do not redesign):
- desktop/src/apps/MessagesApp/MobileMessages.tsx owns the mobile-aware
  toolbar strip (the showToolbar block + the 'Messages' wordmark
  visibility and the title prop). Behaviour, copy and styling are
  identical to the previous in-place JSX; this is a relocation, not
  a redesign.
- desktop/src/apps/MessagesApp/MobileMessages.test.tsx covers the four
  branches: desktop + no channel, mobile + no channel, mobile + channel
  selected (toolbar hidden), desktop + channel selected (toolbar still
  visible), the title override path on both desktop and mobile, and
  the new-channel click forwarding.

Acceptance proof (npm run test, desktop):
- Test Files  422 passed (422)
- Tests       3700 passed (3700)
- npm run build succeeds (vite build).

Behaviour proof: MessagesApp renders and behaves exactly as before --
the only change in the orchestrator is one import line, one JSX block
replaced by a <MobileMessagesToolbar> call, and the showToolbar const
moved into the extracted component. No state, props, refs, callbacks,
effects, className, copy or routing path has been altered.

Docs-Reviewed: README.md does not enumerate per-app entries and the
app-catalog/ tree is unchanged; this commit is a pure file-shape move
inside desktop/src/apps/MessagesApp/ with one new internal component
(MobileMessagesToolbar) and one new internal test file. No app is
added or removed from the user-visible catalogue, no new entry point
is exposed, and no README-level claim becomes inaccurate.
